+++ src/main/java/me/kansio/client/modules/impl/world/ChestStealer.java
@@ -25,7 +25,7 @@ import net.minecraft.potion.PotionEffect;
 public class ChestStealer extends Module {
 
     private BooleanValue checkChest = new BooleanValue("Check Chest", this, true);
-    private NumberValue<Double> delay = new NumberValue<>("Delay", this, 25.0, 0.0, 1000.0, 1.0);
+    private NumberValue<Integer> delay = new NumberValue<>("Delay", this, 25, 0, 1000, 1);
     private Stopwatch delayCounter = new Stopwatch();
 
     public void onEnable() {
